requirements.txt
networkx
fastapi
uvicorn
pydantic-settings
httpx
aiofiles
python-multipart
pytest
